create comp
